feat: rewrite account and friends handling - account store only uses the /user endpoint - avatar is not shown for a non-logged-in user - friend list is server-side - friends store added (derived from account store)
